Extract desktop state storage helpers
